Foem Details input done as seperate component
